hoist unity context to App
we need access to SendMessage in various points in the component tree so
need to have the unity context higher up

in addition, we shouldn't really be redrawing the unity context anyway
so having it at the top of the tree reduces crashes when in dev mode
